requirements.txt
mysql-connector-python
requests
flask
flask-cors
gunicorn
